+++ src/components/global/CheckBox.tsx
@@ -1,6 +1,4 @@
 import './checkmark.css'
-import '../assets/done.svg'
-import { ChangeEventHandler } from 'react';
 
 interface AppState {
   onChange: any;
