+++ src/components/global/IntroSection.tsx
