+++ src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import Footer from "../components/Footer"
+import Footer from "../components/home/Footer"
 
 export const HomePage = () => {
   return (
