+++ src/pages/Profile.tsx
@@ -6,9 +6,9 @@ import pfp from "../assets/pfp.png";
 
 import upload from "../assets/upload_icon.svg";
 import { useState ,useRef } from "react";
-import CheckBox from "../components/CheckBox";
-import TextField, { InputType } from "../components/TextField";
-import Dropdown, { PositionType } from "../components/Dropdown";
+import CheckBox from "../components/global/CheckBox";
+import TextField, { InputType } from "../components/global/TextField";
+import Dropdown, { PositionType } from "../components/global//Dropdown";
 
 const Profile = () => {
   const [inPerson, setInPerson] = useState(false);
